+++ src/feature/alarm/index.ts
@@ -1,3 +1,3 @@
-export { useRegisterTokenMutation } from './api/mutations';
+export { useFcmTokenMutation, useRegisterTokenMutation } from './api/mutations';
 export { StudentAlarmItem } from './ui/StudentAlarmItem';
 export { TrainerAlarmItem } from './ui/TrainerAlarmItem';
