+++ src/feature/auth/index.ts
@@ -1,3 +1,4 @@
+export { useFindIdMutation, useFindPasswordMutation } from './api/mutations';
 export { useInvitationInfoQuery } from './api/queries';
 export { useSignUpFunnel } from './hooks/useSignUpFunnel';
 export { SignInForm } from './ui/SignInForm';
