+++ src/feature/course/const.ts
@@ -6,10 +6,3 @@ export const courseHistoryTypes = {
   RESERVATION: '수업 예약',
   RESERVATION_CANCEL: '수업 예약 취소',
 };
-
-export const pointHistoryTypes = {
-  NO_SHOW: '노쇼',
-  NO_SHOW_CANCEL: '노쇼취소',
-  WORKOUT: '개인운동',
-  DIET: '식단등록',
-};
