+++ src/feature/course/ui/CourseCard.tsx
