+++ src/feature/home/index.ts
@@ -1,0 +1,2 @@
+export { useStudentHomeDataQuery, useTrainerHomeQuery } from './api/queries';
+export type { BestStudent, TrainerHomeInfo } from './model/type';
