+++ src/feature/manage/api/mutations.ts
@@ -1,6 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 
 import { authApi } from '@/entity/auth';
+import { AppendMemberForm, InviteForm } from '@/feature/manage';
 import { BaseError, BaseResponse } from '@/shared/api';
 
 interface EditmemoRequest {
@@ -49,6 +50,65 @@ export const useToggleAlarmStatusMutation = () => {
     mutationFn: async ({ type, status }) => {
       const result = await authApi.patch<BaseResponse<boolean>>(
         `/api/members/v1/alarm/${type}/${status}`
+      );
+      return result.data;
+    },
+  });
+};
+
+type AppendMemberRequest = AppendMemberForm & {
+  memberId: number;
+};
+
+interface AppendMemberResponse {
+  uuid: string;
+}
+
+export const useAppendMemberMutation = () => {
+  return useMutation<BaseResponse<AppendMemberResponse>, BaseError, AppendMemberRequest>({
+    mutationFn: async ({ lessonCnt, memberId }) => {
+      const result = await authApi.post<BaseResponse<AppendMemberResponse>>(
+        `/api/trainers/v1/members/${memberId}`,
+        { lessonCnt }
+      );
+      return result.data;
+    },
+  });
+};
+
+export const useDeleteRefundStudentMutation = () => {
+  return useMutation<BaseResponse<null>, BaseError, number>({
+    mutationFn: async (memberId: number) => {
+      const result = await authApi.delete<BaseResponse<null>>(
+        `/api/trainers/v1/members/${memberId}/refund`
+      );
+      return result.data;
+    },
+  });
+};
+
+export const useDeleteStudentMutation = () => {
+  return useMutation<BaseResponse<null>, BaseError, number>({
+    mutationFn: async (memberId: number) => {
+      const result = await authApi.delete<BaseResponse<null>>(
+        `/api/trainers/v1/members/${memberId}`
+      );
+      return result.data;
+    },
+  });
+};
+
+interface InviteResponse {
+  uuid: string;
+  invitationLink: string;
+}
+
+export const useInviteStudentMutation = () => {
+  return useMutation<BaseResponse<InviteResponse>, BaseError, InviteForm>({
+    mutationFn: async (invitationInfo) => {
+      const result = await authApi.post<BaseResponse<InviteResponse>>(
+        '/api/trainers/v1/invitation',
+        invitationInfo
       );
       return result.data;
     },
