+++ src/feature/manage/index.ts
@@ -1,8 +1,37 @@
+export type {
+  AppendMemberForm,
+  CourseHistoryItem,
+  CourseItem,
+  FindIdRequest,
+  FindIdResponse,
+  FindPasswordRequest,
+  FindPasswordResponse,
+  HistoryType,
+  InviteForm,
+  Member,
+  pointHistoryType,
+  RegisteredStudent,
+  StudentCourse,
+  StudentDetail,
+  StudentPoint,
+  StudentPointItem,
+  StudentRank,
+} from '../manage/model/types';
 export {
+  useAppendMemberMutation,
+  useDeleteRefundStudentMutation,
+  useDeleteStudentMutation,
   useEditMemoMutation,
   useEditNicknameMutation,
+  useInviteStudentMutation,
   useToggleAlarmStatusMutation,
 } from './api/mutations';
+export {
+  useNotRegisteredStudentsQuery,
+  useRegisteredStudentsQuery,
+  useStudentDetailQuery,
+} from './api/queries';
+export { AddStudentDialog } from './ui/AddStudentDialog';
 export {
   CourseSheet,
   CourseSheetContent,
