+++ src/feature/mypage/ui/GymVerificationCode.tsx
