+++ src/feature/point/const.ts
@@ -1,0 +1,6 @@
+export const pointHistoryTypes = {
+  NO_SHOW: '노쇼',
+  NO_SHOW_CANCEL: '노쇼취소',
+  WORKOUT: '개인운동',
+  DIET: '식단등록',
+};
