+++ src/feature/point/index.ts
@@ -1,0 +1,2 @@
+export { useMyPointHistoryQuery, useStudentPointHistoryQuery } from './api/queries';
+export { pointHistoryTypes } from './const';
