+++ src/page/home/index.ts
@@ -1,4 +1,7 @@
-export { type TrainerHomeInfo, useTrainerHomeQuery } from './api/useTrainerHomeQuery';
+export {
+  type TrainerHomeInfo,
+  useTrainerHomeQuery,
+} from '../../feature/home/api/queries';
 export { StudentAlarmPage } from './ui/StudentAlarmPage';
 export { StudentHomePage } from './ui/StudentHomePage';
 export { StudentMyCourseDetailPage } from './ui/StudentMyCourseDetailPage';
