+++ src/page/home/ui/StudentMyPointDetailPage.tsx
@@ -10,7 +10,7 @@ import Link from 'next/link';
 import { useEffect, useState } from 'react';
 import { useInView } from 'react-intersection-observer';
 
-import { pointHistoryTypes, useMyPointHistoryQuery } from '@/feature/member';
+import { pointHistoryTypes, useMyPointHistoryQuery } from '@/feature/point';
 import { IconClose, IconNotification } from '@/shared/assets';
 import { Typography } from '@/shared/mixin';
 import { Card, CardContent, CardHeader } from '@/shared/ui';
