+++ src/page/manage/ui/StudentListPage.tsx
@@ -3,7 +3,7 @@
 import Link from 'next/link';
 
 import { StudentList } from '@/feature/manage';
-import { AddStudentDialog } from '@/feature/member';
+import { AddStudentDialog } from '@/feature/manage';
 import { IconBack } from '@/shared/assets';
 import { Typography } from '@/shared/mixin';
 import { cn } from '@/shared/utils';
